backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose
pdfplumber
python-multipart
openai
google-generativeai
redis
celery
apscheduler
playwright
beautifulsoup4
httpx
cryptography
